Moving xpack helper functions into own package
