+++ metricbeat/helper/xpack/xpack.go
@@ -1,4 +1,4 @@
-package helper
+package xpack
 
 import (
 	"fmt"
