+++ metricbeat/helper/xpack/xpack_test.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package helper
+package xpack
 
 import (
 	"fmt"
